Detection is configurable
A person is found if they exist in enough frames during a period of time. The
time is configurable with --detect-length, the amount of frames required is
configurable with --detect-percentage, the rate at whichfps * detect_length *
detect_percentage this detection happens is configurable with --fps.

Example: --fps=10 --detect-length=1.5 --detect-percentage=66

This means that detection happens over 1.5s and will involve a total of 15
frames. From the time of first detection to 1.5s later, if the person is in
66% of the frames (that is, 10), it triggers.

By default, the camera runs at 10fps. The Detect Length is 1 second. And the
percentage is 75%.